requirements.txt
psycopg2-binary>=2.8,<2.9
werkzeug>0.15
django>=2.0,<3.0
djangorestframework
Pillow>=6.0
django_extensions
requests
pre-commit
django-crispy-forms
django-webpack-loader
django-base-url
django-rest-auth
django-cors-headers
marshmallow
django-constance[database]
redis
dj-database-url
gunicorn[gevent]
djangorestframework-gis
setuptools<58.0.0
pygdal==3.2.2.*
django-extensions
